DIY Code for Met Monitoring Station
This code runs on the Particle dashboard. Data is collected by ThingSpeak. The sensors are SoilWatch 10 soil moisture sensor, Adafruit SHT31 temperature/RH sensor, and an contactless infrared thermometer from Melexis. The code has power control so that the sensors will turn on and off for a given interval of time. The median filter helps smooth the variance of the raw voltage for a soil moisture sensor.